// src/app/model/content.ts
export class Content {
    contentId: any;
    description: string;
      dateCreated: any;
      dateDeleted: string;
      userId: string;
      contentType: string;
      contentRating: string;
      contentAddedBy: string;
      contentLength: number;
      categoryList: string;
      title: string;
      digitalAgency: string;
      contentRepositoryLink: string;
      contentStatus: number;
      deletedUserId: string;
      deletedDate: string
}